add file-position-object and locate-source cache
Allow a file-position object in place of a file-position integer, so
that lines and column information can be recorded in a source object.
Add a cache for `locate-source` as enabled by the `use-cache?`
optional argument, which can avoid compilation times that are
quadratic in the number of `let-values` or `define-values` forms.